+++ pkg/frontend/transport/handler_test.go
@@ -101,35 +101,46 @@ func TestHandler_ServeHTTP(t *testing.T) {
 
 func TestHandler_FailedRoundTrip(t *testing.T) {
 	for _, test := range []struct {
-		name            string
-		cfg             HandlerConfig
-		expectedMetrics int
+		name                string
+		expectedMetrics     int
+		path                string
+		expectQueryParamLog bool
+		queryErr            error
 	}{
 		{
-			name:            "Failed round trip with context cancelled",
-			cfg:             HandlerConfig{QueryStatsEnabled: true},
-			expectedMetrics: 0,
+			name:                "Failed round trip with context cancelled",
+			expectedMetrics:     0,
+			path:                "/api/v1/query?query=up&time=2015-07-01T20:10:51.781Z",
+			expectQueryParamLog: true,
+			queryErr:            context.Canceled,
+		},
+		{
+			name:                "Failed round trip with no query params",
+			expectedMetrics:     0,
+			path:                "/api/v1/query",
+			expectQueryParamLog: false,
+			queryErr:            context.Canceled,
 		},
 	} {
 		t.Run(test.name, func(t *testing.T) {
 			roundTripper := roundTripperFunc(func(req *http.Request) (*http.Response, error) {
-				return nil, context.Canceled
+				return nil, test.queryErr
 			})
 
 			reg := prometheus.NewPedanticRegistry()
 			logs := &concurrency.SyncBuffer{}
 			logger := log.NewLogfmtLogger(logs)
-			handler := NewHandler(test.cfg, roundTripper, logger, reg)
+			handler := NewHandler(HandlerConfig{QueryStatsEnabled: true}, roundTripper, logger, reg)
 
 			ctx := user.InjectOrgID(context.Background(), "12345")
-			req := httptest.NewRequest("GET", "/api/v1/query?query=up&time=2015-07-01T20:10:51.781Z", nil)
+			req := httptest.NewRequest("GET", test.path, nil)
 			req = req.WithContext(ctx)
 			resp := httptest.NewRecorder()
 
 			handler.ServeHTTP(resp, req)
 			require.Equal(t, StatusClientClosedRequest, resp.Code)
 
-			count, _ := promtest.GatherAndCount(
+			count, err := promtest.GatherAndCount(
 				reg,
 				"cortex_query_seconds_total",
 				"cortex_query_fetched_series_total",
@@ -137,7 +148,11 @@ func TestHandler_FailedRoundTrip(t *testing.T) {
 				"cortex_query_fetched_chunks_total",
 			)
 
-			assert.Contains(t, strings.TrimSpace(logs.String()), "sharded_queries", "param_query")
+			assert.NoError(t, err)
+			assert.Contains(t, strings.TrimSpace(logs.String()), "sharded_queries")
+			if test.expectQueryParamLog {
+				assert.Contains(t, strings.TrimSpace(logs.String()), "param_query")
+			}
 			assert.Equal(t, test.expectedMetrics, count)
 		})
 	}